+++ src/commands/info/role.ts
@@ -1,43 +1,9 @@
 import { Command } from 'discord-akairo';
 import { Message, Permissions, Role } from 'discord.js';
+import { getDefaultEmbed } from '../../utils/message';
+import { PERMISSIONS } from '../../utils/permissions';
 import { utc } from 'moment';
 import 'moment-duration-format';
-import { getDefaultEmbed } from '../../utils/message';
-
-interface PermissionsIndex {
-    [key: string]: string;
-}
-
-const PERMISSIONS: PermissionsIndex = {
-    ADMINISTRATOR: 'Administrator',
-    VIEW_AUDIT_LOG: 'View audit log',
-    MANAGE_GUILD: 'Manage server',
-    MANAGE_ROLES: 'Manage roles',
-    MANAGE_CHANNELS: 'Manage channels',
-    KICK_MEMBERS: 'Kick members',
-    BAN_MEMBERS: 'Ban members',
-    CREATE_INSTANT_INVITE: 'Create instant invite',
-    CHANGE_NICKNAME: 'Change nickname',
-    MANAGE_NICKNAMES: 'Manage nicknames',
-    MANAGE_EMOJIS: 'Manage emojis',
-    MANAGE_WEBHOOKS: 'Manage webhooks',
-    VIEW_CHANNEL: 'Read text channels and see voice channels',
-    SEND_MESSAGES: 'Send messages',
-    SEND_TTS_MESSAGES: 'Send TTS messages',
-    MANAGE_MESSAGES: 'Manage messages',
-    EMBED_LINKS: 'Embed links',
-    ATTACH_FILES: 'Attach files',
-    READ_MESSAGE_HISTORY: 'Read message history',
-    MENTION_EVERYONE: 'Mention everyone',
-    USE_EXTERNAL_EMOJIS: 'Use external emojis',
-    ADD_REACTIONS: 'Add reactions',
-    CONNECT: 'Connect',
-    SPEAK: 'Speak',
-    MUTE_MEMBERS: 'Mute members',
-    DEAFEN_MEMBERS: 'Deafen members',
-    MOVE_MEMBERS: 'Move members',
-    USE_VAD: 'Use voice ativity',
-};
 
 export default class RoleInfoCommand extends Command {
     public constructor() {
